Remove ridiculous function limit. Happy day.
The most recent 3.1 firmware update allows 64k functions. Finally. According
to analytics data, it looks like some users are using versions that still
have the 768 limit. For now, I've tweaked the Makefile to enforce a limit
of 768 and removed the note from the README. By the time we run into 768,
we should be able to live with restricting ourselves to updated units that
allow at least 64k.