Make sure Prefs is focused if nothing else is present.
This is an annoying Roku bug, but one that isn't too hard to work around.
Normally when we make all the other rows invisible, it will simply show the
first visible row. But for some reason it messes that up in this scenario,
maybe because we're quickly showing another screen on top of this one?