Switch to mixed-aspect-ratio grids.
And all of the attendant headaches. Oodles of crashes later, it seems that you
can't mark a row as invisible if you might want to make it visible again later.

The home screen and search screens use various dimensions in their rows.
Filtered grids also use the mixed grid to show squares in the filter row
and appropriate dimensions in the rest of the rows.