Improve setting tokens on servers.
Don't blindly set our super token on servers. Only set the super token on
servers from myPlex that don't have their own token. Also, remember tokens
for servers so that we can communicate with them immediately on startup,
without having to wait for the myPlex response that has the tokens.