Support for lite servers.
All of the other clients allow servers to have multiple connections, and then
routinely test those connections to select an active connection. This isn't
quite that, but it's a first step. All of the code that assumes a server has
a single URL is still there. Internally, the server object can have multiple
connections and can test their connections and notify a listener when it finds
an active connection.

That new logic is used to create server objects from (my)Plex XML, and add
each local address as a connection. The home screen data loader then asks
the server to test its connections and kicks off the rest of the server
requests once a valid connection is found.

Though it's still trickier than that, thanks to the grid screen's propensity
to crash when row visibility or focus changes. So we jump through some hoops
to try to avoid hiding rows until we know they're never going to be populated.