Force Roku to show our selected subtitles.
We're not really supposed to do this, we're supposed to respect the user's
global preference (even though 99% of users don't know about that). But that
doesn't actually make sense. Our subtitle selection is richer than that, with
things like respecting default subs, and enabling subs on foreign audio. Plus,
if the subs end up burned in, we wouldn't be respecting the Roku's preference
anyway. So we just ignore the Roku's preference and restore our old behavior.

If anyone complains, there's a hidden preference that can be set with launch
params.

Fixes plexinc/roku-client-public#279.