Always add token to image transcode URLs, fix section icons.
We used to try to be cute and use headers instead of query string args for
screens that only loaded content from a single server. This got much harder
with the mixed-aspect-ratio grid, since we recompute the URL based on the
row style, and some screens don't have a single server. Just put it in the
query string and move on.