Don't necessarily prevent direct streaming for ref frames.
By forcing the transcode/remux decision here, we were making it impossible
for bold/silly users to tweak the Roku.xml profile in PMS and try to direct
stream 9 reference frames.

It may make more sense to have a preference akin to the H.264 level. But
this is a start.