Load/Reload ALL Home Screen rows (mixed-aspect-ratio grid)
The Home Screen should re-load the rows. This will fix hiding, and showing
rows depending on the users preferences. The home screen has a limited number
of rows, so it shouldn't cause any slowdowns.

This is all due to a bug in the Roku firmware that causes the Roku to
potentially reboot if any rows are hidden or shown after the grid has already
been created.