The crash/fix for mixed-aspect-ratio is only needed on the Home Screen.
It seems the crash is also related to the mixed-aspect-ratio grids when the
SetListPosterStyles method is used. The current fix, showing a dialog screen
before hiding/showing a row can be a bit intrusive in library sections when
filtering is disabled. Instead of using that fix, we can exclude calling
SetListPosterStyles when filtering is disabled to inhibit the crash. The trade
off is the fact the mixed-aspect-grid defaults to Portrait. This means
sections like music will use Portrait instead of the Square style. Portrait is
however, the current mode the Official channel is using. It may not be a
problem since anyone who chooses to disabled filtering will probably be
expecting the original functionality.