Add "Connect Plex account" to trial expiration screen.
plexinc/roku-client#43

If the user's trial has expired and they haven't connected a Plex account
yet, add an option to the expiration screen to link an account in addition
to the "Purchase channel" option.

This is better than before, but still not great. Users can't sign up from
the Roku, and they can't purchase a subscription from the Roku. So they
have to solve the problem somewhere else. And we can really only show
the option if the user hasn't connected an account. If a non-PlexPass
account is already connected, the screen is unchanged.