Rework videoCanDirectPlay to place more faith in PMS.
Fixes plexinc/roku-client-public#276
Fixes plexinc/roku-client#34

Instead of doing our own heuristics about stream order and channel count
to make guesses about commentary, just trust that PMS has selected the right
audio stream for us.

That's the easy part, but we still need all sorts of messy logic to decide
if the Roku will actually play the stream that we want it to. We can tell it
which language we want, so if the selected stream is the first of its language
then we should be fine. But we're also fine if the first stream is stereo AAC
and the second is surround AC3, both the same language, and the surround stream
is selected.