Refresh Plex Pass account status more often.
Fixes plexinc/roku-client#44.

We'll now force the status to update after checking the status on the
preferences screen and upon closing the screens that tell the user that
their trial has expired. This should allow users to purchase Plex Pass
while the trial message is on the screen and have the change get picked
up immediately.