Never hide rows on the home screen, never crash.
Fixes plexinc/roku-client#47

The home screen is simply too complicated and the mixed aspect ratio grid
too buggy, it's not worth it to try to hide any rows dynamically. We're smarter
about never creating empty rows in the first place now, which helps a great
deal. If a row is empty and would have been hidden, it will instead show a
placeholder. Selecting that placeholder shows a dialog explaining why there's
no content and reminding the user that they can hide that row in the
preferences if it will never have content.