Set bitrate for direct play audio content metadata.
This theoretically allows Roku to do more accurate buffering math and avoid
stuttering through the first few seconds of high bitrate FLAC. And so it
theoretically fixes plexinc/roku-client#48.

In practice, I still have a file that stutters. I've shared the details with
Roku and we're trying to sort out whether there's anything more I can do or
if there's a demuxer issue that will need to be fixed in the firmware.