Process dialog messages without actually blocking.
Some dialogs want to call their message loop in a blocking way, but we don't
actually want to use a separate message port and block execution. Among other
things, it means that the companion app can't interact with dialogs. So we've
moved the actual innards of the message loop into a function that can be
called from outside the view controller, and the "blocking" dialogs can run
the loop themselves.

Fixes plexinc/roku-client#25