Initial pass at reverting section grid type, plus filters.
Instead of using the hack that is the chunked grid, revert to the traditional
grid used for displaying sections. Add a dummy item in the search row to bring
up the filter options. When there's an active filter, add a row to the grid
for showing the results of the current filter/sort. The last used filters
are remembered per section, so it's possible to set a useful filter and have
it always show up on the main section display. The filtered row can also be
reordered just like any other.

plexinc/roku-client#53