Fix random ordering of items when multi-servers are present.
It's possible due the response time of the servers to have the items on the
global homescreen rows (On Deck, Channels, Recently Added) changing positions
when refreshing the content. To fix, we sort the refreshed content by the
existing server order.

Recently Added has also been added to the refresh routine.