more buggy mixed aspect grid work (inhibit reboots).
There was still a possibly crash/reboot, while it was intermittent and took a
lot of work to reproduce, it happened. This update should hopefully inhibit
most of the reboots. We now try and focus the last selected and visible row
before hiding the said row. Using 999 used to work, but that could cause an
odd phantom/endless scroll that continued to the end of all rows. This will
put the user back into the area they were in before hiding a row.