Cache Plex Pass status in registry.
If plex.tv is down, then we will use the registry to determine Plex Pass
status. The channel already caches a purchase if the Roku store is
down. This will fix anyone trying to use the channel when either their
Internet, Plex.tv and/or the Roku store are down.

closes #61