use a square blank thumb for channels content.
This mainly reflects a change on the iTunes channel as it's the only one
I can find that is sometimes missing a thumbUrl in then xml
response. Either way, channels are never portrait, so it should make
sense for any other channel (com.plexapp.plugins)

I wanted to change the default displayMode to photo-fit, but that may be
a little too intrusive for now.