backwards compatibility for mixed/non-mixed grids
 Mixed / Non-Mixed counterpart
 * portrait  /  flat-portrait
 * landscape /  flat-16x9 (flat-landscape needs border asset)
 * square    /  flat-square

createGridScreen can be passed a third option if any of the above do not
fit. I.E. for music/photos we use flat-movie since it's shorter than the
flat-portrait.

We will also verify the style is valid (square,portrait or landscape) if
we are creating mixed grid, otherwise a reboot will happen.